requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
jinja2>=3.1.3
playwright>=1.43.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
